init commit

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -9,33 +9,19 @@ use Illuminate\Support\Facades\Storage;
 
 class CategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
-        $categories=Category::get();
+        $categories=Category::has('products')->get();
         return view('admin.categories.index',compact('categories'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
         return view('admin.categories.form');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
         $request->validate([
@@ -50,12 +36,6 @@ class CategoryController extends Controller
         return redirect()->route('categories.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         if(!$category = Category::find($id)){
@@ -65,12 +45,7 @@ class CategoryController extends Controller
         return view('admin.categories.show',compact('category'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit($id)
     {
         if(!$category = Category::find($id)){
@@ -79,13 +54,7 @@ class CategoryController extends Controller
         return view('admin.categories.form',compact('category'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
         if(!$category = Category::find($id)){
@@ -115,7 +84,9 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-
+        if(!$category = Category::find($id)){
+            return abort(404);
+        }
         $category->delete();
         return redirect()->route('categories.index');
     }
